requirements/test3.txt
nose
django-nose
nose-progressive
coverage
